moved mock data to a different file and wrote a skeleton of the add users test
