+++ src/index.test.ts
@@ -1,7 +1,17 @@
 //write tests here
+import UserAPI from "./index"
+import {
+  mockUsers,
+  userObject,
+  updatingUserObject,
+  updatingUserObjectError,
+  updatingUserObjectErrorWithId,
+} from "./mockdata"
 
-describe('Tests will go here!', () => {
-  it('should pass', () => {
+describe("Testing User API", () => {
+  it("should add a user to and throw an error if no name is provided", () => {
+    const users = new UserAPI(mockUsers)
+    users.addUser({})
     expect(true).toBeTruthy()
   })
 })
